fix(clippy): resolve all clippy warnings across workspace
- Remove redundant #[must_use] on Result-returning and iterator methods
- Use direct `as u32` casts for count conversions (counts never exceed u32)
- Add type alias ChannelIdExtractor for complex boxed closure type
- Use explicit MessageDeliveryStatus::default() instead of Default::default()
- Remove redundant variable reassignments
- Use struct initialization with ..Default::default() pattern
- Add #[allow] for intentional patterns (NonZeroU8::new(1), recursive methods)
- Inline format args and remove unnecessary clones

+++ crates/aura-sync/src/services/sync.rs
@@ -366,8 +366,8 @@ impl SyncService {
         SyncServiceHealth {
             status,
             active_sessions: session_stats.active_sessions as u32,
-            tracked_peers: peer_stats.total_tracked as u32,
-            available_peers: peer_stats.available_peers as u32,
+            tracked_peers: peer_stats.total_tracked,
+            available_peers: peer_stats.available_peers,
             last_sync,
             uptime,
         }
